+++ sites/cnfpt/src/backend/cnfpt/settings.py
@@ -363,6 +363,7 @@ class Base(StyleguideMixin, DRFMixin, RichieCoursesConfigurationMixin, Configura
         "djangocms_picture",
         "djangocms_text_ckeditor",
         "djangocms_video",
+        "django_check_seo",
         "cms",
         "menus",
         "sekizai",
@@ -461,7 +462,7 @@ class Base(StyleguideMixin, DRFMixin, RichieCoursesConfigurationMixin, Configura
     }
 
     # Demo
-    RICHIE_DEMO_SITE_DOMAIN = "localhost:8080"
+    RICHIE_DEMO_SITE_DOMAIN = "localhost:8000"
     RICHIE_DEMO_FIXTURES_DIR = os.path.join(BASE_DIR, "base", "fixtures")
 
     # Elasticsearch
@@ -595,6 +596,12 @@ class Base(StyleguideMixin, DRFMixin, RichieCoursesConfigurationMixin, Configura
     # in the page tree admin view
     CMS_PAGETREE_DESCENDANTS_LIMIT = 80
 
+    # - Django CMS Check SEO
+    # Excludes all elements that are not related to the page content
+    DJANGO_CHECK_SEO_EXCLUDE_CONTENT = (
+        "body > svg, #main-menu, .body-footer, .body-mentions"
+    )
+
     # pylint: disable=invalid-name
     @property
     def ENVIRONMENT(self):
@@ -660,6 +667,9 @@ class Development(Base):
             "LOCATION": "dev-cache",
         }
     }
+
+    # Django Check SEO
+    DJANGO_CHECK_SEO_FORCE_HTTP = True
 
 
 class Test(Base):