+++ sites/funcampus/src/backend/funcampus/settings.py
@@ -295,6 +295,7 @@ class Base(StyleguideMixin, DRFMixin, RichieCoursesConfigurationMixin, Configura
         "djangocms_picture",
         "djangocms_text_ckeditor",
         "djangocms_video",
+        "django_check_seo",
         "cms",
         "menus",
         "sekizai",
@@ -449,7 +450,7 @@ class Base(StyleguideMixin, DRFMixin, RichieCoursesConfigurationMixin, Configura
     }
 
     # Demo
-    RICHIE_DEMO_SITE_DOMAIN = "localhost:8080"
+    RICHIE_DEMO_SITE_DOMAIN = "localhost:8000"
     RICHIE_DEMO_FIXTURES_DIR = os.path.join(BASE_DIR, "base", "fixtures")
 
     # Elasticsearch
@@ -552,6 +553,12 @@ class Base(StyleguideMixin, DRFMixin, RichieCoursesConfigurationMixin, Configura
     # in the page tree admin view
     CMS_PAGETREE_DESCENDANTS_LIMIT = 80
 
+    # - Django CMS Check SEO
+    # Excludes all elements that are not related to the page content
+    DJANGO_CHECK_SEO_EXCLUDE_CONTENT = (
+        "body > svg, #main-menu, .body-footer, .body-mentions"
+    )
+
     # pylint: disable=invalid-name
     @property
     def ENVIRONMENT(self):
@@ -616,6 +623,9 @@ class Development(Base):
             "BACKEND": "django.core.cache.backends.locmem.LocMemCache",
         }
     }
+
+    # Django Check SEO
+    DJANGO_CHECK_SEO_FORCE_HTTP = True
 
     RICHIE_DEMO_NB_OBJECTS = {"licences": 0}
     # Note: This override can be simplified once a merge update is added in richie for PAGES_INFO