+++ sites/funmooc/src/backend/funmooc/settings.py
@@ -361,6 +361,7 @@ class Base(StyleguideMixin, DRFMixin, RichieCoursesConfigurationMixin, Configura
         "djangocms_picture",
         "djangocms_text_ckeditor",
         "djangocms_video",
+        "django_check_seo",
         "cms",
         "menus",
         "sekizai",
@@ -574,7 +575,7 @@ class Base(StyleguideMixin, DRFMixin, RichieCoursesConfigurationMixin, Configura
     )
 
     # Demo
-    RICHIE_DEMO_SITE_DOMAIN = "localhost:8080"
+    RICHIE_DEMO_SITE_DOMAIN = "localhost:8000"
     RICHIE_DEMO_FIXTURES_DIR = os.path.join(BASE_DIR, "base", "fixtures")
 
     # Elasticsearch
@@ -675,6 +676,12 @@ class Base(StyleguideMixin, DRFMixin, RichieCoursesConfigurationMixin, Configura
     # in the page tree admin view
     CMS_PAGETREE_DESCENDANTS_LIMIT = 80
 
+    # - Django CMS Check SEO
+    # Excludes all elements that are not related to the page content
+    DJANGO_CHECK_SEO_EXCLUDE_CONTENT = (
+        "#tarteaucitronRoot, body > svg, #main-menu, .body-footer, .body-mentions"
+    )
+
     # pylint: disable=invalid-name
     @property
     def ENVIRONMENT(self):
@@ -735,6 +742,9 @@ class Development(Base):
     ALLOWED_HOSTS = ["*"]
 
     CACHES = {"default": {"BACKEND": "django.core.cache.backends.locmem.LocMemCache"}}
+
+    # Django Check SEO
+    DJANGO_CHECK_SEO_FORCE_HTTP = True
 
 
 class Test(Base):